+++ attributes/default.rb
@@ -112,6 +112,7 @@ default['openstack']['identity']['ldap']['page_size'] = 0
 default['openstack']['identity']['ldap']['alias_dereferencing'] = 'default'
 default['openstack']['identity']['ldap']['query_scope'] = 'one'
 default['openstack']['identity']['ldap']['use_tls'] = false
+# If both tls_cacertfile and tls_cacertdir are set, then the cacertfile takes precedence and tls_cacertdir is not used
 default['openstack']['identity']['ldap']['tls_cacertfile'] = nil
 default['openstack']['identity']['ldap']['tls_cacertdir'] = nil
 default['openstack']['identity']['ldap']['tls_req_cert'] = 'demand'
@@ -125,7 +126,6 @@ default['openstack']['identity']['ldap']['user_name_attribute'] = 'sn'
 default['openstack']['identity']['ldap']['user_mail_attribute'] = 'email'
 default['openstack']['identity']['ldap']['user_pass_attribute'] = 'userPassword'
 default['openstack']['identity']['ldap']['user_enabled_attribute'] = 'enabled'
-default['openstack']['identity']['ldap']['user_domain_id_attribute'] = 'businessCategory'
 default['openstack']['identity']['ldap']['user_enabled_mask'] = 0
 default['openstack']['identity']['ldap']['user_enabled_default'] = 'true'
 default['openstack']['identity']['ldap']['user_attribute_ignore'] = 'tenant_id,tenants'
@@ -136,6 +136,7 @@ default['openstack']['identity']['ldap']['user_enabled_emulation'] = false
 default['openstack']['identity']['ldap']['user_enabled_emulation_dn'] = nil
 
 # LDAP backend tenant related settings
+# NOTE! tenant_xxxx has been deprecated in Juno, please use project_xxxx below
 default['openstack']['identity']['ldap']['tenant_tree_dn'] = nil
 default['openstack']['identity']['ldap']['tenant_filter'] = nil
 default['openstack']['identity']['ldap']['tenant_objectclass'] = 'groupOfNames'
@@ -151,6 +152,22 @@ default['openstack']['identity']['ldap']['tenant_allow_update'] = true
 default['openstack']['identity']['ldap']['tenant_allow_delete'] = true
 default['openstack']['identity']['ldap']['tenant_enabled_emulation'] = false
 default['openstack']['identity']['ldap']['tenant_enabled_emulation_dn'] = nil
+
+default['openstack']['identity']['ldap']['project_tree_dn'] = node['openstack']['identity']['ldap']['tenant_tree_dn']
+default['openstack']['identity']['ldap']['project_filter'] = node['openstack']['identity']['ldap']['tenant_filter']
+default['openstack']['identity']['ldap']['project_objectclass'] = node['openstack']['identity']['ldap']['tenant_objectclass']
+default['openstack']['identity']['ldap']['project_id_attribute'] = node['openstack']['identity']['ldap']['tenant_id_attribute']
+default['openstack']['identity']['ldap']['project_member_attribute'] = node['openstack']['identity']['ldap']['tenant_member_attribute']
+default['openstack']['identity']['ldap']['project_name_attribute'] = node['openstack']['identity']['ldap']['tenant_name_attribute']
+default['openstack']['identity']['ldap']['project_desc_attribute'] = node['openstack']['identity']['ldap']['tenant_desc_attribute']
+default['openstack']['identity']['ldap']['project_enabled_attribute'] = node['openstack']['identity']['ldap']['tenant_enabled_attribute']
+default['openstack']['identity']['ldap']['project_domain_id_attribute'] = node['openstack']['identity']['ldap']['tenant_domain_id_attribute']
+default['openstack']['identity']['ldap']['project_attribute_ignore'] = node['openstack']['identity']['ldap']['tenant_attribute_ignore']
+default['openstack']['identity']['ldap']['project_allow_create'] = node['openstack']['identity']['ldap']['tenant_allow_create']
+default['openstack']['identity']['ldap']['project_allow_update'] = node['openstack']['identity']['ldap']['tenant_allow_update']
+default['openstack']['identity']['ldap']['project_allow_delete'] = node['openstack']['identity']['ldap']['tenant_allow_delete']
+default['openstack']['identity']['ldap']['project_enabled_emulation'] = node['openstack']['identity']['ldap']['tenant_enabled_emulation']
+default['openstack']['identity']['ldap']['project_enabled_emulation_dn'] = node['openstack']['identity']['ldap']['tenant_enabled_emulation_dn']
 
 # LDAP backend role related settings
 default['openstack']['identity']['ldap']['role_tree_dn'] = nil
@@ -172,7 +189,6 @@ default['openstack']['identity']['ldap']['group_id_attribute'] = 'cn'
 default['openstack']['identity']['ldap']['group_name_attribute'] = 'ou'
 default['openstack']['identity']['ldap']['group_member_attribute'] = 'member'
 default['openstack']['identity']['ldap']['group_desc_attribute'] = 'description'
-default['openstack']['identity']['ldap']['group_domain_id_attribute'] = 'businessCategory'
 default['openstack']['identity']['ldap']['group_attribute_ignore'] = nil
 default['openstack']['identity']['ldap']['group_allow_create'] = true
 default['openstack']['identity']['ldap']['group_allow_update'] = true